Grid progress
This grid is tricky. I want to make it so you can click maps to add them to the grid in the first open slot, but also drag maps into a specific slot and between slots. Hold shift to insert instead of replacing. Resize may push some maps off, or bring some in that were selected that there weren't room for. I can't figure out how to get an event for IsSelected changing on any map.