The most insane code ever
Okay, so the idea is to make it so imports are asynchronous. That is, you click confirm all, and the images come in over time.

First step is to make OnConfirmed provide a list, and also make the list lazy, because creating the objects upfront is expensive because auto-scaling needs to access the size of the image.

Then we need to think about undoing and redo. When we undo, we want to stop processing maps, and remove any added so far. When we redo, we want to re-add the removed maps, then continue processing.

So basically, the action kickstarts a task, then passes context to the undo. The undo cancels the task, undoes its work, then passes context back to the action for redo.