Create command line application
Ok so, originally the plan was to make the normal exe work as a command line program if you give it the right arguments. But apparently you have to call AttachConsole which does this weird thing where it prints your path twice when you call it, and it was just weird. The other option is switching to a console-type application, but that makes the console show up next to the form. So third option is split the project. All of the model stuff was moved over, keeping the UI stuff in the main project. This means I also had to remove stuff like ImageSource properties, instead making a converter that does a WeakReference cache creating the sources, and stuff like a JavaInventory wrapper to keep the display stuff in the UI.
Anyway then it got annoying because when you publish the project, the command line one isn't self-contained even though it's supposed to be, so it doesn't run properly. So I have to publish it separately to the same folder, but I do it in the build step so it's fine. Also turned on ReadyToRun which apparently makes it launch faster but makes the file size bigger. Also I stop disposing bedrock worlds because it makes it crash if you're loading maps during that time. So I guess we just leave them open forever now.